The point where you realize a complete rethink is necessary
